Create new component for favorite spots preferences
